+++ src/index.js
@@ -4,7 +4,8 @@ import { createBrowserHistory } from "history";
 import { Router, Route, Switch, Redirect } from "react-router-dom";
 
 // core components
-import Dashboard from "layouts/Dashboard/Dashboard.jsx";
+import Admin from "layouts/Admin/Admin.jsx";
+import RTL from "layouts/RTL/RTL.jsx";
 
 import "assets/css/material-dashboard-react.css?v=1.5.0";
 
@@ -13,7 +14,8 @@ const hist = createBrowserHistory();
 ReactDOM.render(
   <Router history={hist}>
     <Switch>
-      <Route path="/admin" component={Dashboard} />
+      <Route path="/admin" component={Admin} />
+      <Route path="/rtl" component={RTL} />
       <Redirect from="/" to="/admin/dashboard" />
     </Switch>
   </Router>,