+++ src/views/RTLPage/RTLPage.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+class RTLPage extends React.Component {
+  render() {
+    return <div>Hey</div>;
+  }
+}
+
+export default RTLPage;